refactor: Store money as DECIMAL and BigDecimal across model and DTOs.

+++ src/main/java/ru/yandex/practicum/mymarket/dto/OrderView.java
@@ -1,10 +1,11 @@
 package ru.yandex.practicum.mymarket.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderView(
         long id,
         List<OrderItemView> items,
-        long totalSum
+        BigDecimal totalSum
 ) {
 }
